Auto-restore scratch buffer in autosave minor mode
Fixes #10.

When desktop buffers are restored, their minor modes are also
restored, including peristent-scratch-autosave-mode. If
persistent-scratch buffer is not restored before the autosave timer
expires (e.g., a bug in the init.el script which results in
persistent-scratch-setup-default never being called), then the
scratch file will be overwritten with an empty scratch buffer.

By adding once-per session automatic scratch restoration to the
autosave minor mode, the buffer is guaranteed to be restored even in
this corner case, avoiding the possibility of overwriting the scratch file with
an empty scratch buffer.